some UI tweaking + notes for later
