+++ src/lib/integrations/daoStack/arc/types.ts
@@ -1,5 +1,10 @@
 // types exported by this integration
 
+// TODO: clean up the types, define the paradigm for 3 layers of types (UI, LIB, 3rd party)
+// - VotingMachine|(Param | Configuration) is a great example of why defining this seperation
+//   will be useful. We don't want to bleed UI data into the state (descriptions, display names, etc),
+//   but we also want to keep them consistent between tools (GUI, CLI, etc).
+
 export type VotingMachineConfiguration = {
   typeName: string
   params: { [paramName: string]: string | number }
